requirements.txt
# Ana AI Kütüphaneleri
torch>=2.0.0
diffusers>=0.18.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0

# Web Arayüzü
gradio>=3.32.0

# Görsel İşleme
pillow>=9.0.0
opencv-python>=4.7.0
moviepy>=1.0.0

# GPU ve Sistem İzleme
nvidia-ml-py3>=7.352.0
psutil>=5.9.0

# Yardımcı Kütüphaneler
numpy>=1.24.0
requests>=2.28.0
tqdm>=4.64.0
